Rename GnuArFile.cs to ArFile
We now also support BSD and SystemV formatted Ar files in the same parser.
